Remove Collection class
This duplicate its code in GraphNode and GraphEdge
