+++ src/GraphNode/GraphEdge.php
@@ -29,7 +29,7 @@ use Facebook\Exception\SDKException;
 /**
  * @package Facebook
  */
-class GraphEdge extends Collection
+class GraphEdge implements \ArrayAccess, \Countable, \IteratorAggregate
 {
     /**
      * @var Request the original request that generated this data
@@ -52,6 +52,13 @@ class GraphEdge extends Collection
     protected $subclassName;
 
     /**
+     * The items contained in the collection.
+     *
+     * @var array
+     */
+    protected $items = [];
+
+    /**
      * Init this collection of GraphNode's.
      *
      * @param Request     $request            the original request that generated this data
@@ -66,8 +73,169 @@ class GraphEdge extends Collection
         $this->metaData = $metaData;
         $this->parentEdgeEndpoint = $parentEdgeEndpoint;
         $this->subclassName = $subclassName;
+        $this->items = $data;
+    }
 
-        parent::__construct($data);
+    /**
+     * Gets the value of a field from the Graph node.
+     *
+     * @param string $name    the field to retrieve
+     * @param mixed  $default the default to return if the field doesn't exist
+     *
+     * @return mixed
+     */
+    public function getField($name, $default = null)
+    {
+        if (isset($this->items[$name])) {
+            return $this->items[$name];
+        }
+
+        return $default;
+    }
+
+    /**
+     * Returns a list of all fields set on the object.
+     *
+     * @return array
+     */
+    public function getFieldNames()
+    {
+        return array_keys($this->items);
+    }
+
+    /**
+     * Get all of the items in the collection.
+     *
+     * @return array
+     */
+    public function all()
+    {
+        return $this->items;
+    }
+
+    /**
+     * Get the collection of items as a plain array.
+     *
+     * @return array
+     */
+    public function asArray()
+    {
+        return array_map(function ($value) {
+            if ($value instanceof GraphNode || $value instanceof GraphEdge) {
+                return $value->asArray();
+            }
+
+            return $value;
+        }, $this->items);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function map(\Closure $callback)
+    {
+        return new static(
+            $this->request,
+            array_map($callback, $this->items, array_keys($this->items)),
+            $this->metaData,
+            $this->parentEdgeEndpoint,
+            $this->subclassName
+        );
+    }
+
+    /**
+     * Get the collection of items as JSON.
+     *
+     * @param int $options
+     *
+     * @return string
+     */
+    public function asJson($options = 0)
+    {
+        return json_encode($this->asArray(), $options);
+    }
+
+    /**
+     * Count the number of items in the collection.
+     *
+     * @return int
+     */
+    public function count()
+    {
+        return count($this->items);
+    }
+
+    /**
+     * Get an iterator for the items.
+     *
+     * @return \ArrayIterator
+     */
+    public function getIterator()
+    {
+        return new \ArrayIterator($this->items);
+    }
+
+    /**
+     * Determine if an item exists at an offset.
+     *
+     * @param mixed $key
+     *
+     * @return bool
+     */
+    public function offsetExists($key)
+    {
+        return array_key_exists($key, $this->items);
+    }
+
+    /**
+     * Get an item at a given offset.
+     *
+     * @param mixed $key
+     *
+     * @return mixed
+     */
+    public function offsetGet($key)
+    {
+        return $this->items[$key];
+    }
+
+    /**
+     * Set the item at a given offset.
+     *
+     * @param mixed $key
+     * @param mixed $value
+     *
+     * @return void
+     */
+    public function offsetSet($key, $value)
+    {
+        if (is_null($key)) {
+            $this->items[] = $value;
+        } else {
+            $this->items[$key] = $value;
+        }
+    }
+
+    /**
+     * Unset the item at a given offset.
+     *
+     * @param string $key
+     *
+     * @return void
+     */
+    public function offsetUnset($key)
+    {
+        unset($this->items[$key]);
+    }
+
+    /**
+     * Convert the collection to its string representation.
+     *
+     * @return string
+     */
+    public function __toString()
+    {
+        return $this->asJson();
     }
 
     /**
@@ -231,19 +399,5 @@ class GraphEdge extends Collection
         }
 
         return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function map(\Closure $callback)
-    {
-        return new static(
-            $this->request,
-            array_map($callback, $this->items, array_keys($this->items)),
-            $this->metaData,
-            $this->parentEdgeEndpoint,
-            $this->subclassName
-        );
     }
 }
